+++ packages/framer-motion/src/utils/use-in-view.ts
@@ -29,7 +29,7 @@ export function useInView(
         }
 
         return inView(ref.current, onEnter, options)
-    }, [root, ref, margin, once])
+    }, [root, ref, margin, once, amount])
 
     return isInView
 }